Update 0.6
- Turrets detect enemies and removes them

- Bug with the bullets 👍 (Most of them probably are bcz they cant get the target)

+++ LA QUE NO ES DE M17/AndroidDefense/Assets/Scripts/Scriptables/Enemies/EnemyInfo.cs
@@ -7,7 +7,7 @@ public class EnemyInfo : ScriptableObject
 {
     public int hp;
     public int damage;
-    public int velocity;
+    public float velocity;
     public int gold;
     public Color color; //MAS ADELANTE ALOMEJOR INTENTO CAMBIARLO POR UN SPRITE ANIMADO
 }
